Rename Dto project -> Shared
