+++ DragaliaAPI.Photon.Plugin/DragaliaAPI.Photon.Plugin.csproj
@@ -16,7 +16,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<ProjectReference Include="..\DragaliaAPI.Photon.Dto\DragaliaAPI.Photon.Dto.csproj" />
+	  <ProjectReference Include="..\DragaliaAPI.Photon.Shared\DragaliaAPI.Photon.Shared.csproj" />
 	</ItemGroup>
 
 	<ItemGroup>
