+++ DragaliaAPI.Photon.Plugin/GluonPlugin.cs
@@ -4,14 +4,13 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
-using DragaliaAPI.Photon.Dto;
-using DragaliaAPI.Photon.Dto.Game;
-using DragaliaAPI.Photon.Dto.Requests;
 using DragaliaAPI.Photon.Plugin.Constants;
-using DragaliaAPI.Photon.Plugin.Extensions;
 using DragaliaAPI.Photon.Plugin.Helpers;
 using DragaliaAPI.Photon.Plugin.Models;
 using DragaliaAPI.Photon.Plugin.Models.Events;
+using DragaliaAPI.Photon.Shared;
+using DragaliaAPI.Photon.Shared.Models;
+using DragaliaAPI.Photon.Shared.Requests;
 using MessagePack;
 using Newtonsoft.Json;
 using Photon.Hive.Plugin;
