+++ DragaliaAPI.Photon.Shared/Requests/GameModifyRequest.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using DragaliaAPI.Photon.Dto.Game;
-
-namespace DragaliaAPI.Photon.Dto.Requests
+﻿namespace DragaliaAPI.Photon.Shared.Requests
 {
     /// <summary>
     /// A request to modify a game -- either join, leave, or close.
