+++ DragaliaAPI.Test/Helpers/BaasRequestHelperTest.cs
@@ -2,8 +2,8 @@
 using DragaliaAPI.Models;
 using DragaliaAPI.Models.Generated;
 using DragaliaAPI.Models.Options;
+using DragaliaAPI.Services.Api;
 using DragaliaAPI.Services.Exceptions;
-using DragaliaAPI.Services.Helpers;
 using DragaliaAPI.Shared.Json;
 using Microsoft.Extensions.Caching.Distributed;
 using Microsoft.Extensions.Caching.Memory;
