+++ DragaliaAPI.Test/Services/AuthServiceTest.cs
@@ -4,8 +4,8 @@ using DragaliaAPI.Models;
 using DragaliaAPI.Models.Generated;
 using DragaliaAPI.Models.Options;
 using DragaliaAPI.Services;
+using DragaliaAPI.Services.Api;
 using DragaliaAPI.Services.Exceptions;
-using DragaliaAPI.Services.Helpers;
 using DragaliaAPI.Shared;
 using DragaliaAPI.Test.Utils;
 using Microsoft.Extensions.Logging;
