+++ DragaliaAPI/Program.cs
@@ -7,7 +7,6 @@ using DragaliaAPI.Models.Options;
 using DragaliaAPI.Services;
 using DragaliaAPI.Services.Api;
 using DragaliaAPI.Services.Health;
-using DragaliaAPI.Services.Helpers;
 using DragaliaAPI.Services.Photon;
 using DragaliaAPI.Shared;
 using DragaliaAPI.Shared.Json;
