+++ DragaliaAPI/Services/Api/IBaasApi.cs
@@ -1,7 +1,7 @@
 ﻿using DragaliaAPI.Models.Generated;
 using Microsoft.IdentityModel.Tokens;
 
-namespace DragaliaAPI.Services.Helpers;
+namespace DragaliaAPI.Services.Api;
 
 public interface IBaasApi
 {
