+++ DragaliaAPI/Services/Api/PhotonStateApi.cs
@@ -1,4 +1,4 @@
-﻿using DragaliaAPI.Photon.Dto.Game;
+﻿using DragaliaAPI.Photon.Shared.Models;
 using Microsoft.AspNetCore.Http.Extensions;
 using Microsoft.AspNetCore.WebUtilities;
